Fix some typing
