+++ src/db/schema.ts
@@ -1,4 +1,4 @@
-import { sql } from "drizzle-orm";
+import { InferSelectModel, sql } from "drizzle-orm";
 import {
   pgTable,
   integer,
@@ -9,16 +9,49 @@ import {
   bigint,
 } from "drizzle-orm/pg-core";
 
+const specialties = [
+  "Bipolar",
+  "LGBTQ",
+  "Medication/Prescribing",
+  "Suicide History/Attempts",
+  "General Mental Health (anxiety, depression, stress, grief, life transitions)",
+  "Men's issues",
+  "Relationship Issues (family, friends, couple, etc)",
+  "Trauma & PTSD",
+  "Personality disorders",
+  "Personal growth",
+  "Substance use/abuse",
+  "Pediatrics",
+  "Women's issues (post-partum, infertility, family planning)",
+  "Chronic pain",
+  "Weight loss & nutrition",
+  "Eating disorders",
+  "Diabetic Diet and nutrition",
+  "Coaching (leadership, career, academic and wellness)",
+  "Life coaching",
+  "Obsessive-compulsive disorders",
+  "Neuropsychological evaluations & testing (ADHD testing)",
+  "Attention and Hyperactivity (ADHD)",
+  "Sleep issues",
+  "Schizophrenia and psychotic disorders",
+  "Learning disorders",
+  "Domestic abuse",
+] as const;
+
+type Specialty = (typeof specialties)[number];
+
 const advocates = pgTable("advocates", {
   id: serial("id").primaryKey(),
   firstName: text("first_name").notNull(),
   lastName: text("last_name").notNull(),
   city: text("city").notNull(),
   degree: text("degree").notNull(),
-  specialties: jsonb("payload").default([]).notNull(),
+  specialties: jsonb("payload").default([]).$type<Specialty[]>().notNull(),
   yearsOfExperience: integer("years_of_experience").notNull(),
   phoneNumber: bigint("phone_number", { mode: "number" }).notNull(),
   createdAt: timestamp("created_at").default(sql`CURRENT_TIMESTAMP`),
 });
 
-export { advocates };
+type Advocate = InferSelectModel<typeof advocates>;
+
+export { advocates, specialties, type Advocate };
